+++ changelog.md
@@ -1,3 +1,9 @@
+# 1.1.0 (2018/03/20)
+* Enable caching of genomic reference data
+* Enable processing of multiple GTC files or directories for improved performance
+* Address issue with incorrect assignment of reference allele for indels
+* Added script to build reference FASTA files for human
+
 # 1.0.1 (2017/10/27)
 * Eliminated extraneous header in column